backend/requirements.txt
# API Framework
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.30

# AI & NLP (V0.3+ for better Pydantic V2 support)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=1.0.3
google-generativeai>=0.5.2
openai>=1.50.0
tiktoken>=0.7.0

# Document Processing
pypdf>=4.2.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
loguru>=0.7.2
typing-extensions>=4.12.0
tenacity>=8.5.0
